add: table assets
