+++ src/widgets/winnersTable/ui/customColumns/customPlaceCol.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import {HStack} from "shared/ui/stack";
+
+export const FirstColumnRenderer: React.FC<{ value: string }> = ({ value }) => {
+    return (
+        <HStack className="place-col">
+            {value}
+        </HStack>
+    );
+};
